refactor(trading-signals): fold performance metrics into trading and rename statistics to math
Performance evaluation is part of the trading workflow — plan, time,
signal, evaluate — so the metrics join the trading helpers. The numeric
category is honestly named math: extremes and percentage arithmetic were
never statistics.
