+++ packages/trading-signals/src/momentum/ER/ER.ts
@@ -1,7 +1,7 @@
 import {TradingSignal, TrendIndicatorSeries} from '../../base/Indicator.js';
 import type {HighLowClose} from '../../base/Candle.type.js';
-import {getMaximum} from '../../util/statistics/getMaximum.js';
-import {getMinimum} from '../../util/statistics/getMinimum.js';
+import {getMaximum} from '../../util/math/getMaximum.js';
+import {getMinimum} from '../../util/math/getMinimum.js';
 import {pushUpdate} from '../../util/array/pushUpdate.js';
 
 export type ERThresholds = {
