+++ packages/trading-signals/src/trend/FRAMA/FRAMA.ts
@@ -1,5 +1,5 @@
-import {getMaximum} from '../../util/statistics/getMaximum.js';
-import {getMinimum} from '../../util/statistics/getMinimum.js';
+import {getMaximum} from '../../util/math/getMaximum.js';
+import {getMinimum} from '../../util/math/getMinimum.js';
 import {pushUpdate} from '../../util/array/pushUpdate.js';
 import {MovingAverage} from '../MA/MovingAverage.js';
