+++ packages/trading-signals/src/trend/VHF/VHF.ts
@@ -1,6 +1,6 @@
 import {IndicatorSeries} from '../../base/Indicator.js';
-import {getMaximum} from '../../util/statistics/getMaximum.js';
-import {getMinimum} from '../../util/statistics/getMinimum.js';
+import {getMaximum} from '../../util/math/getMaximum.js';
+import {getMinimum} from '../../util/math/getMinimum.js';
 import {pushUpdate} from '../../util/array/pushUpdate.js';
 
 /**
