+++ packages/trading-signals/src/util/index.ts
@@ -5,25 +5,25 @@ export * from './candle/getTrueRange.js';
 export * from './candle/getTypicalPrice.js';
 export * from './candle/getWeightedClose.js';
 export * from './candle/mergeCandles.js';
-export * from './performance/getCAGR.js';
-export * from './performance/getLogReturns.js';
-export * from './performance/getMaxDrawdown.js';
-export * from './performance/getReturns.js';
-export * from './performance/getSharpeRatio.js';
-export * from './statistics/addPercentageChange.js';
-export * from './statistics/getAverage.js';
-export * from './statistics/getCorrelation.js';
-export * from './statistics/getLinearRegression.js';
-export * from './statistics/getMaximum.js';
-export * from './statistics/getMedian.js';
-export * from './statistics/getMinimum.js';
-export * from './statistics/getPercentageChange.js';
-export * from './statistics/getQuartile.js';
-export * from './statistics/getShare.js';
-export * from './statistics/getStandardDeviation.js';
-export * from './statistics/getStreaks.js';
-export * from './statistics/getZScore.js';
+export * from './math/addPercentageChange.js';
+export * from './math/getAverage.js';
+export * from './math/getCorrelation.js';
+export * from './math/getLinearRegression.js';
+export * from './math/getMaximum.js';
+export * from './math/getMedian.js';
+export * from './math/getMinimum.js';
+export * from './math/getPercentageChange.js';
+export * from './math/getQuartile.js';
+export * from './math/getShare.js';
+export * from './math/getStandardDeviation.js';
+export * from './math/getStreaks.js';
+export * from './math/getZScore.js';
+export * from './trading/getCAGR.js';
 export * from './trading/getGrid.js';
+export * from './trading/getLogReturns.js';
+export * from './trading/getMaxDrawdown.js';
+export * from './trading/getReturns.js';
+export * from './trading/getSharpeRatio.js';
 export * from './trading/getWeekday.js';
 export * from './trading/hasCrossedOver.js';
 export * from './trading/hasCrossedUnder.js';
