+++ packages/trading-signals/src/util/math/addPercentageChange.test.ts
