+++ packages/trading-signals/src/util/math/addPercentageChange.ts
