+++ packages/trading-signals/src/util/math/getAverage.test.ts
