+++ packages/trading-signals/src/util/math/getAverage.ts
