+++ packages/trading-signals/src/util/math/getCorrelation.test.ts
