+++ packages/trading-signals/src/util/math/getCorrelation.ts
