+++ packages/trading-signals/src/util/math/getLinearRegression.test.ts
