+++ packages/trading-signals/src/util/math/getLinearRegression.ts
