+++ packages/trading-signals/src/util/math/getMaximum.test.ts
