+++ packages/trading-signals/src/util/math/getMaximum.ts
