+++ packages/trading-signals/src/util/math/getMedian.ts
