+++ packages/trading-signals/src/util/math/getMinimum.test.ts
