+++ packages/trading-signals/src/util/math/getMinimum.ts
