+++ packages/trading-signals/src/util/math/getPercentageChange.test.ts
