+++ packages/trading-signals/src/util/math/getPercentageChange.ts
