+++ packages/trading-signals/src/util/math/getQuartile.test.ts
