+++ packages/trading-signals/src/util/math/getQuartile.ts
