+++ packages/trading-signals/src/util/math/getShare.test.ts
