+++ packages/trading-signals/src/util/math/getShare.ts
