+++ packages/trading-signals/src/util/math/getStandardDeviation.ts
