+++ packages/trading-signals/src/util/math/getStreaks.test.ts
