+++ packages/trading-signals/src/util/math/getStreaks.ts
