+++ packages/trading-signals/src/util/trading/getCAGR.test.ts
