+++ packages/trading-signals/src/util/trading/getCAGR.ts
