+++ packages/trading-signals/src/util/trading/getLogReturns.test.ts
