+++ packages/trading-signals/src/util/trading/getLogReturns.ts
