+++ packages/trading-signals/src/util/trading/getMaxDrawdown.test.ts
