+++ packages/trading-signals/src/util/trading/getMaxDrawdown.ts
