+++ packages/trading-signals/src/util/trading/getReturns.test.ts
