+++ packages/trading-signals/src/util/trading/getReturns.ts
@@ -1,4 +1,4 @@
-import {getPercentageChange} from '../statistics/getPercentageChange.js';
+import {getPercentageChange} from '../math/getPercentageChange.js';
 
 /**
  * Converts a price series into per-bar simple returns, in percent (`25` means "+25%"). Returns are
