+++ packages/trading-signals/src/util/trading/getSharpeRatio.test.ts
